JWrappingLabel no longer breaks the wizards Windows installer won't build if wget.exe or 7zS.sfx are not found Linux launcher now passes params with spaces in them correctly to the client DoReports now also sources ${HOME}/.bashrc to get the correct classpath
git-svn-id: http://svn.dmdirc.com/trunk@3140 00569f92-eb28-0410-84fd-f71c24880f43

+++ installer/windows/makeInstallerWindows.sh
@@ -472,6 +472,11 @@ if [ ! -e "wget.exe" ]; then
 	wget "http://users.ugent.be/~bpuype/cgi-bin/fetch.pl?dl=wget/wget.exe"
 fi;
 
+if [ ! -e "wget.exe" ]; then
+	echo "wget.exe not found, unable to continue."
+	exit 1;
+fi;
+
 FILES="${FILES} wget.exe"
 
 compress $FILES
@@ -494,6 +499,11 @@ if [ ! -e "7zS.sfx" ]; then
 	tar -jxvf 7zextra.tar.bz2 7zS.sfx
 	rm 7zextra.tar.bz2
 fi
+
+if [ ! -e "7zS.sfx" ]; then
+	echo "7zS.sfx not found, unable to continue."
+	exit 1;
+fi;
 
 echo "Creating .exe"
 cat 7zS.sfx 7zip.conf "${INTNAME}" > "${RUNNAME}"
